Update README.md
Agriculture in every country plays a prominent role in its economy and employment. But it is noticed that there is a common problem faced amongst the farmers while farming, that they don't choose the right crop based on their soil requirements. Due to which, product's quality will get degraded and farmers face a severe setback into their productivity.